hackcoin/requirements.txt
APScheduler==3.2.0
click==6.7
crypto==1.4.1
ed25519==1.4
Flask==0.11.1
Flask-Assets==0.12
Flask-Redis==0.3.0
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.6
MarkupSafe==1.0
Naked==0.1.31
psycopg2==2.7.1
PyJWT==1.4.2
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.7.0
shellescape==3.4.1
six==1.11.0
tzlocal==1.5.1
webassets==0.12.1
Werkzeug==0.14.1
